requirements.txt
Flask==2.0.1
gunicorn==20.1.0
theblockchainapi==0.0.25
numpy==1.21.4
urllib3==1.26.7
Pillow==8.4.0
cloudinary==1.28.0